+++ src/test/java/jpa/domain/SectionTest.java
@@ -1,14 +1,14 @@
-package jpa.repository;
+package jpa.domain;
 
-import jpa.domain.Line;
-import jpa.domain.Section;
-import jpa.domain.Station;
+import jpa.repository.LineRepository;
+import jpa.repository.StationRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 @DataJpaTest
 public class SectionTest {
